Add a function called game
the game function basically made a 5 playRound function with round count
and score count. Here's how it works - initialize round and score for
player and computer then inside the for loop. round will be equal to the
counter of for loop. Get the user input from prompt then the return
message value will be slice from 0 to the first occurance of "!". Use
that in the if else to determine which side scored. Lasly alert
everything